use size_t

+++ utils/array.h
@@ -10,16 +10,16 @@ namespace giacomo {
 template <typename Type>
 struct array {
     Type* data  = nullptr;
-    int   count = 0;
+    size_t   count = 0;
 
-    inline const Type& operator[](int i) const { return data[i]; }
-    inline Type&       operator[](int i) { return data[i]; }
+    inline const Type& operator[](size_t i) const { return data[i]; }
+    inline Type&       operator[](size_t i) { return data[i]; }
     inline const Type& back() const { return data[count - 1]; }
     inline Type&       back() { return data[count - 1]; }
     inline void        add(const Type& val) { data[count++] = val; }
 
     inline int  size() const { return count; }
-    inline void resize(int new_size) { count = new_size; }
+    inline void resize(size_t new_size) { count = new_size; }
     inline void push_back(const Type& val) { add(val); }
 
     array() {
@@ -28,21 +28,21 @@ struct array {
     }
 
     // array(const Type* p, int c) : data(p), count(c) {}
-    array(Type* p, int c) : data(p), count(c) {}
+    array(Type* p, size_t c) : data(p), count(c) {}
 
-    inline void insert(const Type& element, int index) {
-        for (int i = count; i > index; i--) {
+    inline void insert(const Type& element, size_t index) {
+        for (size_t i = count; i > index; i--) {
             data[i] = data[i - 1];
         }
         data[index] = element;
         count += 1;
     }
 
-    inline void insert(const array<Type>& arr, int index) {
-        for (int i = count; i > index; i--) {
+    inline void insert(const array<Type>& arr, size_t index) {
+        for (size_t i = count; i > index; i--) {
             data[i] = data[i - arr.count];
         }
-        for (int i = 0; i < arr.count; ++i) {
+        for (size_t i = 0; i < arr.count; ++i) {
             data[index + i] = arr[i];
         }
         count += arr.count;
@@ -50,29 +50,29 @@ struct array {
 
     inline void operator+=(const array<Type>& arr) { insert(arr, count); }
 
-    inline void remove(int index) {
-        for (int i = index; i < count - 1; ++i) {
+    inline void remove(size_t index) {
+        for (size_t i = index; i < count - 1; ++i) {
             data[i] = data[i + 1];
         }
         count -= 1;
     }
 
     void operator=(const std::initializer_list<Type>& list) {
-        count = (int)list.size();
-        int i = 0;
+        count = list.size();
+        size_t i = 0;
         for (auto& v : list) {
             data[i++] = v;
         }
     }
 
-    array<Type> slice(int from, int to) {
+    array<Type> slice(size_t from, size_t to) {
         assert(from >= 0 and from <= count);
         assert(to >= 0 and to <= count);
         assert(from <= to);
         return {data + from, to - from};
     }
 
-    const array<const Type> slice(int from, int to) const {
+    const array<const Type> slice(size_t from, size_t to) const {
         assert(from >= 0 and from <= count);
         assert(to >= 0 and to <= count);
         assert(from <= to);
@@ -82,7 +82,7 @@ struct array {
     // iterators
     struct const_iterator {
         const Type* data;
-        int         i;
+        size_t         i;
 
         const_iterator& operator++() {
             i += 1;
@@ -98,7 +98,7 @@ struct array {
 
     struct mutable_iterator {
         Type* data;
-        int   i;
+        size_t   i;
 
         mutable_iterator& operator++() {
             i += 1;
@@ -132,14 +132,14 @@ inline void fill(array<Type>& arr, const Type& val) {
 // fill array with values returned by a function f: (int) -> Type
 template <typename Type, typename Function>
 inline void fill(array<Type>& arr, Function f) {
-    for (int i = 0; i < arr.size(); ++i) arr[i] = f(i);
+    for (size_t i = 0; i < arr.size(); ++i) arr[i] = f(i);
 }
 
 // copy content of array to another array
 template <typename Type>
 inline void copy_to(const array<Type>& from, array<Type>& to) {
     assert(from.count <= to.count);
-    for (int i = 0; i < from.count; ++i) to[i] = from[i];
+    for (size_t i = 0; i < from.count; ++i) to[i] = from[i];
     to.count = from.count;
 }
 
@@ -147,16 +147,16 @@ inline void copy_to(const array<Type>& from, array<Type>& to) {
 template <typename Type>
 inline void copy_to(const array<array<Type>>& from, array<array<Type>>& to) {
     assert(from.count <= to.count);
-    for (int i = 0; i < from.count; ++i) copy_to(from[i], to[i]);
+    for (size_t i = 0; i < from.count; ++i) copy_to(from[i], to[i]);
     to.count = from.count;
 }
 
 // get index of fisrt occurrence of some value
 template <typename Type>
-inline int find(const array<Type>& arr, const Type& value) {
-    for (int i = 0; i < arr.count; ++i)
+inline size_t find(const array<Type>& arr, const Type& value) {
+    for (size_t i = 0; i < arr.count; ++i)
         if (arr[i] == value) return i;
-    return -1;
+    return (size_t)-1;
 }
 
 // check if array contains some value
@@ -168,16 +168,16 @@ inline bool contains(const array<Type>& arr, const Type& value) {
 template <typename Type>
 inline bool operator==(const array<Type>& a, const array<Type>& b) {
     if (a.count != b.count) return false;
-    for (int i = 0; i < a.count; i++) {
+    for (size_t i = 0; i < a.count; i++) {
         if (a[i] != b[i]) return false;
     }
     return true;
 }
 
-inline int argmin(const array<int>& arr) {
+inline size_t argmin(const array<int>& arr) {
     assert(arr.count > 0);
-    int min_index = 0;
-    for (int i = 1; i < arr.count; ++i) {
+    size_t min_index = 0;
+    for (size_t i = 1; i < arr.count; ++i) {
         if (arr[i] < arr[min_index]) {
             min_index = i;
         }
@@ -185,10 +185,10 @@ inline int argmin(const array<int>& arr) {
     return min_index;
 }
 
-inline int argmax(const array<int>& arr) {
+inline size_t argmax(const array<int>& arr) {
     assert(arr.count > 0);
-    int max_index = 0;
-    for (int i = 1; i < arr.count; ++i) {
+    size_t max_index = 0;
+    for (size_t i = 1; i < arr.count; ++i) {
         if (arr[i] > arr[max_index]) {
             max_index = i;
         }
@@ -197,7 +197,7 @@ inline int argmax(const array<int>& arr) {
 }
 
 inline void shuffle(array<int>& arr) {
-    for (int i = arr.count - 1; i > 0; --i) {
+    for (size_t i = arr.count - 1; i > 0; --i) {
         int j   = rand() % (i + 1);
         int tmp = arr[i];
         arr[i]  = arr[j];
